% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiqt.R
\name{fiqt}
\alias{fiqt}
\title{A function for FDR Inverse Quantile Transformation (FIQT)}
\usage{
fiqt(z = z, min.p = 10^-300)
}
\arguments{
\item{z}{a vector of association Z-scores}

\item{min.p}{minimum p-value admitted (to avoid zero p-values/adjusted p-values which give troubles with inverse cdf)
very large Z-scores corresponding to min.p (i.e., z > 37) are not adjusted, as their bias is essentially zero}
}
\value{
mu.z a vector of FIQT (winner's curse) adjusted association Z-scores
}
\description{
This function computes the FIQT (winner's curse) adjusted assocation Z-scores
}
